Create a reusable set of CalmRecordGenerators
So we can create CalmRecord instances in other test classes.  Note that
this changes the behaviour in a subtle way.  If you look at the example
in the comment, the ordering of values is significant.

Given:

    fields = ("Place" -> "London", "Place" -> "Paris", "Date" -> "2020")

Previously:

    data = Map("Place" -> List("Paris", "London"), "Date" -> "2020")

Now:

    data = Map("Place" -> List("London", "Paris"), "Date" -> "2020")

This means the ordering of values in `data` will match the ordering in
`fields`, rather than being the reverse.  I don't know if this has any
effect on tests, but it should avoid weird debugging in future.