Add a global "max_capacity" variable
If you set this, the max capacity of every service will be *at most* whatever you set
this variable to be.

It's not *exactly* this value -- some services can only run at a certain parallelism,
and we don't want to exceed that.  e.g. the image inferrer can only go so fast before
it overwhelms Loris.  For those services, we use the min of (service max, global max).

The primary use case for this is allowing us to set max_capacity = 0, flipping a stack
into read-only mode in a way that allows us to resume later and not lose any SQS work.
(Within the limits of SQS queue expiry.)