Show a breakdown of workTypes in the API diff tool
When you run the diff tool, you get a table of index statistics.
Something like:

                      prod        staging
    Visible       1,198,480      1,206,132      7,652 ▲
    Redirected      685,612        688,281      2,669 ▲
    Invisible     1,235,123      1,224,821    -10,302 ▼
    Deleted          85,093         85,088         -5 ▼
    TOTAL         3,204,308      3,204,322         14 ▲

If the number of Works in the API has changed significantly, this helps
you work out (1) why and (2) if that's something you expect.

In the example above, I know I've recently fixed a bug that was causing
Works to be wrongly marked as Invisible.  These numbers seem plausible.
But if I wasn't expecting that to happen, these numbers would give me a
clue that there's something worth investigating in this index before I
promote it to prod.