Allow adding extra RDS instances in the pipeline stack
The ID minter throughput is limited by the number of ID minter instances
we have available.  It's useful to add new instances during a reindex, so
the ID minter doesn't run out of connections -- but previously this
meant editing the infra/critical stack, which we want to avoid doing on
a regular basis.

This change allows us to add/remove *extra* RDS instances from the pipeline
stack, without affecting the core cluster resources.