Wait for the WorkGraphStore.put to complete in this test
Logging from CI shows that this test is flaky, and sometimes it doesn't
try to log the component ID 'ABC'.  That would make sense if it doesn't
get that component from the store -- if this future hasn't completed yet.

I'm hoping that waiting for the future to complete will make this test
stop being flaky in CI.