If the location is "in above", try to guess the locationType
Some Sierra locations have names like "bound in above" or
"closed in above", which is completely unhelpful for mapping to
our fixed vocabulary of location types.

We assume that "in above" refers to a location on one of the
other items on this bib, so if we can guess that unambiguously,
use that as the location for this item.

This gets the proportion of locations we can map into the new
model from 95% to 99%.