Handle an HTTP 413 Request Too Large error
This occurs when we try to index more documents than Elasticsearch can
handle in a single request.  Thus, we divide and conquer -- split the
list of documents in two, and index both halves separately.  Eventually
we will either index everything, or find the single document that can't
be indexed (and everything else will index successfully).

In theory this could cause recursion issues; unfortunately I couldn't
see a good way to make this tail-recursive.

In practice I don't think that'll be too much of an issue -- we only see
413s a handful of times during a reindex, and I think a single recursion
will fix it.  If not, we'll find out quickly!