Include some tracing in the logs for Elasticsearch HTTP 413 retries
Previously:

    WARN HTTP 413 from Elasticsearch (100 documents); trying smaller slices

Now:

    WARN HTTP 413 from Elasticsearch (100 documents); trying smaller slices (trace 4a0481d54621da7387b28b32889c91be => b1d8a8c818ac49a2fe7f69c099637742 / 8acc3a4f0ca57fd33d5dddcdd814e2c5})
    INFO Received both results from HTTP 413 retry (trace 4a0481d54621da7387b28b32889c91be => b1d8a8c818ac49a2fe7f69c099637742 / 8acc3a4f0ca57fd33d5dddcdd814e2c5})

I have a sneaky feeling that retries aren't actually happening, and
eventually the merger gives up -- which would be bad!  Given we're
already in an unusual scenario, including some extra tracing information
isn't too annoying.

The trace IDs are mostly meaningless, and just need to be consistent
within a single message event.  I want something that's readable and easy
to find in the logs.  You should be able to see if an HTTP 413 got split
in two and never actually returned.