Give the METS adapter permission to talk to the Internet
Previously it was failing at startup, with an error that sounded like it
couldn't talk to Secrets Manager.  It was the same error in this
Stack Overflow question: https://stackoverflow.com/q/61265108/1558022

AFAICT, the previous version had no way to contact the public Internet --
including both Secrets Manager and the storage service.  I'm not quite
sure how that happened, but it seems to be working now.