Holdings should only have a single location
We create the location on a holdings record from fixed field 40, which
only occurs once on a Sierra record.  That means we can simplify the
holdings model -- rather than an arbitrary list, instead we make it a
singular location.

Note that some holdings instances won't have a location -- e.g. if it's
a Sierra location code we can't map to a location type.  That's why this
is an Option[Location] rather than a required field.