Fix a broken inference manager test
The problem is that the keys of `images` are instances of `CanonicalId`,
so `req.contains(_)` will always return False -- it only has a
`.contains()` method for Strings.  This means there's no response
when the inference manager tries to get the inferrer results.

Casting the IDs to strings before making this comparison should
help, and the error messages should make this more obvious to
the next person who has a similar issue.

Errors should never pass silently, etc.