We only care about there being zero *identified* items on an e-video
It's possible for e-video Works to pick up unidentified items from
MARC field 856, which confuses the merger and causes it to attach the
IIIF manifest to the wrong Work.  By filtering only for *identified*
items (i.e. those that came directly from Sierra), the manifest gets
attached to the correct work.