Handle a Sierra record with no "name" in the source language
Previously we assumed that the language on a Sierra bib record would
always have a `code` and a `name`, e.g.

    {"code": "tgl", "name": "Tagalog"}

but today's pipeline has some records with just a code, e.g.

    {"code": "tgl"}

The `name` is mostly irrelevant for our transformer – we look up the
code against our own copy of the MARC language code list, and only if
it's unrecognised do we fall back to the name on the source record.
As such, its absence definitely shouldn't cause a record to fail (=DLQ).

This patch modifies the transformer to tolerate the absence of a `name`
on Sierra languages.