Set an explicit time on results sent from the matcher
This is the first part of a change to how matcher versioning works.
This patch just adds the time in the matcher, but it gets ignored by the
versioner.  I'll deploy this first, let any messages in the old format
get processed by the merger, and update the merger to read this value
separately.