Index all the fields on a SourceIdentifier
In the requesting service, we want to be able to find items with a
specific SourceIdentifier.  Unfortunately, right now we can only match
on the value, which isn't unique, e.g.

    value     | identifierType       | ontologyType
    ----------+----------------------+--------------
    b12345678 | sierra-record-number | Work
    b12345678 | mets                 | Work

is totally legal, but without this patch there's
no way to query Elasticsearch and find just one.

This patch changes the index config so we'll index all three fields, and
will allow us to start matching against them.