Allow overriding the license on a Miro record
We need to change the license of some Miro works,  where the source data
is incorrect. This PR allows the transformer to take a list of license
changes which override the source data.

We don't want to make the list of license changes public, so instead we
have the transformer fetch it at startup from S3.

It expects to receive a JSON file of the form:

    {
        "A0000001": {"id": "cc-by"},
        "A0000002": {"id": "cc-by-nc"}
    }

which becomes a Scala `Map[String, License]`.

Constructing this file and passing the appropriate Terraform config will
come separately.