Put access terms from 506 in a note, not an access condition
The access terms in 506 apply to every item on a bib -- for example, if this
bib contains information that's under data protection legislation.

An item can contain additional information that might affect the access of
that particular copy -- for example, if it's too fragile to be produced.

Previously we'd copy the access terms from the bib to every item on the work,
which means they're duplicated if the work has multiple items -- making it
harder to tell which terms are universal, and which terms are for a single item.

This patch moves the 506 to a note, rather than copying it to every item.
This means it will be displayed on /works at the top level rather than having
to be squeezed onto every individual item.  We still copy the status from
our pick list down onto individual items (for now), but not the terms.

Additional notes:

-   Using the logic in SierraNotes means we surface a couple of subfields
    we were previously ignoring, including subfields ǂc and ǂu, both of
    which are used on real records.

    e.g. subfield ǂc
    https://search.wellcomelibrary.org/iii/encore/record/C__Rb3043980?lang=eng

    > Restricted: currently undergoing conservation. Please contact
    > collections@wellcomecollection.org for further information.

    e.g. subfield ǂu
    https://search.wellcomelibrary.org/iii/encore/record/C__Rb3227583?lang=eng

    > restricted access: http://purl.org/coar/access_right/c_16ec.
    > online access with authorization.

-   In theory, we've lost the "to" date from subfield ǂg.

    In practice, this subfield has never been populated in our Sierra catalogue,
    so we don't lose anything by removing this logic.