Apply a strict mapping for varfields
This will avoid Elasticsearch making a bad guess for the field in a way
that makes it impossible to index future data.  In particular, it looks
like the current index has one of these fields mapped as "date", so
anything that can't be coerced to a date is being rejected.

This also marks several fields as keywords, which should reduce the
pressure on the cluster when we're indexing.