Create an item title from the copyNo if no volume is specified
This is to expand the number of items without titles, so that users can
distinguish between a long list of items.

Also, in some bibs (e.g. b10796290), the cataloguers specifically refer
to the numbered copies in the notes -- but there's no way for a user to know
which item they mean without putting the copy number in the title.