Change the behaviour of SierraItemAccess
We're a lot pickier about what we call terms and a note.  The terms
describe what you can do with the thing if you get hold of it
(e.g. copyright restrictions), the note just generally tells you
about the thing.

We also add support for the item-level note, when the note is for
helping readers distinguish between items rather than telling them
how they access it.