Display item-level shelfmarks like 12345i.1
It's possible for a bib to have a reference number like 12345i, and then
different parts of the item to have individual shelfmarks like .1, .2,
and so on.  These are sometimes referred to in the notes, and help
readers distinguish between them.

So in very specific cases, we'll let the shelfmark back through when it:

* Has a common prefix with the reference number
* Isn't the same as the reference number
* Isn't a comma-separated list of i-numbers