Don't warn if the Sierra language code is just whitespace
There are warnings in the Sierra transformer logs like:

    WARN  w.p.t.s.t.SierraLanguages$ - 3182204: Unrecognised primary language: SierraSourceLanguage(   ,None)

This warning is meant to help us spot malformed language codes or data
errors that need correcting, but we know whitespace strings are fine --
they're used for records that don't have a language (e.g. paintings).

The previous mechanism was a bit clever and tried to use a custom decoder;
instead let's handle this just before we log the warning.