Don't require empty WorkData on deleted works
We can hard-code this as empty data, because deleted works will never be
merged or combined with anything else -- this is what we already do on
redirected works.

This is done with an eye towards more radical changes to WorkData,
in particular making "title" compulsory in some future revision.

Because Circe ignores fields in the JSON that it doesn't care about,
this should be fine to drop into an existing pipeline.  Additionally,
this might make reindexes a bit cheaper!  We have a lot of deleted works,
and not having to send an empty `WorkData` JSON blob back and forth will
reduce the amount of traffic we have to send over the VPC endpoint.