Cap the ID minter at 1 task while reindexing
When we're reindexing, we add extra RDS capacity so we can run lots of
ID minter instances.  When we're not reindexing, we only run a single
task in each service to avoid overwhelming RDS and Elasticsearch.

Unfortunately, we were using the var value(=15) instead of the
local (if_reindexing ? 15 : 1), so it was trying to run 15 ID minters.
This quickly falls over because the RDS cluster can't keep up, oops.