Share the index in WorksIndexConfigTest
These tests work by creating dozens of different instances of Work and
trying to store them in the index.  The way we're doing this is very
inefficient -- we were creating a new ES index per Work, and only
looking to see if the index contained *a* Work.

This new approach shares the index, and indexes Works under a known ID,
so we only have to create one index per test.

Where previously an individual test took ~13s locally, now the entire
suite passes in the same time.