Ensure the METS adapter behaves idempotently
Because DynamoDB only supports second-level precision of Instant values,
if we get a bag with a millisecond-level createdDate, we can't store it
if we receive the notification twice -- it looks different to the already
stored bag.

We used to handle this by catching a VersionAlreadyExistsError, but that
swallows errors where DynamoDB has recorded different information.  This
patch tweaks the equality methods on the associated classes so we'll only
allow the double-write if the data in DynamoDB matches what we're trying
to write (by leveraging the idempotence of VersionedStore.put()).