Ditch the WorkIdentifier class; simplify matcher-merger messaging
Currently the matcher leaks information about the matched graph,
e.g. if we have a graph

    A --> B

where we've seen A but not B, we'll send both nodes to the merger --
which has to realise it should ignore B.

This patch modifies the matcher so it only sends the nodes that have
corresponding Works, so the merger can retrieve every Work it's told
about.  This lets us bin one of the too-many classes in the matcher/merger.