Use the new status/opacmsg for "restricted"
Something weird has happened in Sierra so that if a user requests a
restricted item, the item becomes "open" when it returns to the closed
stores.  This is clearly suboptimal.

To fix this, LS&S have changed how restricted items are recorded in Sierra.
Quoting an email from Louise Simon:

    Opacmsg – 'c' Restricted
    I’ve created a new Opacmsg code = 'c' Restricted (NB 'r' was already
    in use for Reading Room).

    Item Status – '6' Restricted
    This code has been deleted.

    I’ve updated all the restricted items so that Status = '-' Available
    and Opacmsg = 'c' Restricted.

This patch applies that change so our code, so restricted items will
appear as such on wc.org/works.