Don't declare separate types for each instance of Note
Creating the implicit JSON encoders/decoders for Note is surprisingly
expensive (~10% of implicit resolution) – because we have a lot of
different subtypes to create encoders/decoders for.

We don't really need the separate types, so I want to try removing them
and switching to something closer to what we do in the API, and see if
it helps the compilation times of internal model.

This also means we can change the Note types in the pipeline without
having to make a corresponding change in the API repo.