Fix a bug where METS/Sierra works weren't being merged correctly
In particular, if you sent the works in this order:

1. The suppressed Sierra e-bib
2. The METS work
3. The unsuppressed Sierra e-bib

We wouldn't create a Sierra->METS link in step 2, so we wouldn't retrieve
the METS work on step 3.

This is a side-effect of us attempting to be clever about how we put
suppressed works in the store.

Now we give the same "component ID" (now a misnomer) to all the works
that are written together, so if two works were written together they'll
always be retrieved together.  We then have to rejig the rest of the
matcher a bit so that individual components are written together, but
overall it's simpler so hopefully we won't have more bugs like this
in future.