Copy identifiers when a physical bib has an unrelated merge candidate
In particular, we want to handle the case where:

    Sierra e-bib
        ↓
    Sierra physical bib
        ↓
    Miro work

Previously we weren't finding the e-bib in "indFirstLinkedDigitisedSierraWorkFor",
because we'd go down the wrong branch if there were *any* merge
candidates, even if they weren't the ones we were looking for.
This meant we weren't putting any identifiers on the merged work.