Provide more info around these "impossible" ID minter errors
The ID minter makes case insensitive database queries, so if you look up
"METS b12345678" the database will happily return "METS B12345678", even
though that's not what you asked for -- and then the ID minter balks.

There's a specific case that happens relatively often (which is to say,
twice) where somebody changes the b number in a METS file.  We can't fix
that automatically, but we can provide more context for the next person
to hit this error.