Add a helper for creating FixedField values
This by itself doesn't do much except reduce boilerplate, but it opens
the door to refactoring these into GivenWhenThen-style tests, where
we put an And() clause in the shared helper, e.g.

```scala
And(s"fixed field $label = '$value' / '$display')
```

which gets us verbose tests without lots of repetition.

We can discuss whether to do that in the patch that adds it -- even if
we don't go down that route, I think this is a useful change.