Speed up the script for reindexing the visible Works
Previously we were batching in two places:

* We'd get a batch of 100 IDs from Elasticsearch
* We'd split those into 10 batches of 10 to send to SNS in PublishBatch

This created a sawtooth-like set of work: we'd have to publish 10 batches
before we could start working on the next batch of 100.  This is the
failure mode described here:
https://alexwlchan.net/2019/10/adventures-with-concurrent-futures/#running-a-large-fixed-number-of-tasks

This changes the script to work on a continuous stream of works, which
improves the concurrency and takes the script from ~25m to ~2m.