Add a way to turn down unused ES clusters
Our Elastic Cloud bill has been climbing a little, and one reason is
that running an API-ready cluster costs ~$30/day.  This gives us a way
to turn it down in a non-destructive way.

This weekend alone will save ~$100 from not running the broken clusters
hot.  They can *probably* be safely deleted… but this means we can turn
them down if the other devs aren't around to check.