Be more tolerant of empty <textLang> nodes in TEI
When we wrote the original TEI transformer, we did warnings-as-errors to
help flush out issues.  Now it's live, an empty node is a bit annoying
but we know what to do with it, and removing it from the TEI may be
wrong -- e.g. in Tamil 42, where it's part of a partially completed
template:

https://github.com/wellcomecollection/wellcome-collection-tei/blob/0a90dc7843dd76054292c1943e6228c0375b1bc2/Tamil/Tamil_42.xml#L82-L84