Don't merge Sierra physical bibs with the digitised AV items
This is extending a rule for video to cover audio and other AV also.
I've added a regression test – I suspect the merger tests could do with
a bit of a spring clean, as they overlap/repeat in a bunch of places, but
that's a bigger piece of work than I want to do right now.