Fix the lookup of bib IDs on order records
On order records, the linked bibs are supplied as a list of URLs to the
resources in the Sierra API, and not a list of individual bib numbers.
This patch updates the reader to extract the bib IDs from the URLs,
which should fix the issue where order records aren't being linked
correctly.