Adds the stale PR workflow (#2593)
This change implements the stale PR template.

## Who is this for?

Developers who want good visibility of work in progress and a process for preventing proliferation of unmerged PRs.

## What is it doing for them?

Closing old PRs that haven't been merged in 30 days. Things that hang about for longer than that should be smaller PRs, released behind a toggle, or be Issues from which other PRs hang.