Use union type in Concept id, instead of base class
The type declaration should be Identifiable | Identified | Unidentifiable instead of BaseIdentify, otherwise the source_identifier and other_identifiers fields will be ignored. Pydantic needs to know it has to choose either Identified or Unidentifiable when loading in the raw data, otherwise it will just load everything into BaseIdentify and discard all fields which don't exist on this class.

Co-Authored-By: Štěpán Brychta <stepan.brychta@me.com>